backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
mistralai==1.0.0
faiss-cpu==1.7.4
numpy==1.26.2
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.2
python-jose==3.3.0
